add trainer and model_api
